Add models and helpers for multi tenancy
This is intended to be a replacement for Apartment using row level security. It has helpers for enabling row level security, and methods for switching tenants that can replace the existing apartment methods fairly seamlessly. It also has a module that can be used to set whether a model is tenanted or not, and a helper to check for whether a model is tenanted.